+++ packages/contracts/contracts/test/invariant/actors/CommitmentTreeManagerHandler.sol
@@ -19,7 +19,7 @@ contract CommitmentTreeManagerHandler is Test {
 
     // ______PUBLIC______
     TestCommitmentTreeManager public commitmentTreeManager;
-    address public SUBTREE_BATCH_FILLER;
+    address public subtreeBatchFiller;
 
     uint256 public ghost_joinSplitLeafCount = 0;
     uint256 public ghost_refundNotesLeafCount = 0;
@@ -43,7 +43,7 @@ contract CommitmentTreeManagerHandler is Test {
         address _subtreeBatchFiller
     ) {
         commitmentTreeManager = _commitmentTreeManager;
-        SUBTREE_BATCH_FILLER = _subtreeBatchFiller;
+        subtreeBatchFiller = _subtreeBatchFiller;
     }
 
     modifier trackCall(bytes32 key) {
@@ -115,7 +115,7 @@ contract CommitmentTreeManagerHandler is Test {
         uint256 leavesLeft = TreeUtils.BATCH_SIZE -
             commitmentTreeManager.currentBatchLen();
         if (leavesLeft != TreeUtils.BATCH_SIZE) {
-            vm.prank(SUBTREE_BATCH_FILLER);
+            vm.prank(subtreeBatchFiller);
             commitmentTreeManager.fillBatchWithZeros();
             ghost_fillBatchWithZerosLeafCount += leavesLeft;
         } else {